requirements.txt
torch
einops
basicsr
torchvision
torchmetrics
streamlit
gdown
pillow
matplotlib
natsort